Flyttade biljettyps väljaren till egen komponent
