+++ WebReact/src/views/TheaterView.jsx
@@ -3,6 +3,7 @@ import { Container, Row, Col, Button, Form, InputGroup } from 'react-bootstrap';
 import { get, post } from '../utilsAndHooks/rest';
 import { Link, useParams, useOutletContext } from "react-router-dom";
 import ShowSeats from "../components/ShowSeats";
+import ShowTicketType from "../components/ShowTicketType";
 import { useNavigate } from 'react-router-dom';
 import createBookingJson from "../utilsAndHooks/createBookingJson";
 
@@ -148,52 +149,6 @@ const TheaterView = () => {
         seatStatusFeed.connect();
     }
 
-    const increaseTicketCount = (category) => {
-        setTickets((prevTickets) => {
-            const updatedTickets = { ...prevTickets };
-            switch (category) {
-                case 'barn':
-                    if (updatedTickets.ordinary > 0) {
-                        updatedTickets.ordinary -= 1;
-                        updatedTickets.child += 1;
-                    }
-                    break;
-                case 'pensionar':
-                    if (updatedTickets.ordinary > 0) {
-                        updatedTickets.ordinary -= 1;
-                        updatedTickets.pensioner += 1;
-                    }
-                    break;
-                default:
-                    break;
-            }
-            return updatedTickets;
-        });
-    };
-
-    const decreaseTicketCount = (category) => {
-        setTickets((prevTickets) => {
-            const updatedTickets = { ...prevTickets };
-            switch (category) {
-                case 'barn':
-                    if (updatedTickets.child > 0) {
-                        updatedTickets.ordinary += 1;
-                        updatedTickets.child -= 1;
-                    }
-                    break;
-                case 'pensionar':
-                    if (updatedTickets.pensioner > 0) {
-                        updatedTickets.ordinary += 1;
-                        updatedTickets.pensioner -= 1;
-                    }
-                    break;
-                default:
-                    break;
-            }
-            return updatedTickets;
-        });
-    };
-
     //När man klickar på ett säte 
     const seatClicked = (seatId) => {
         const updatedSeats = [...seats];
@@ -257,37 +212,8 @@ const TheaterView = () => {
 
             <ShowSeats {...{ seats, theater, seatClicked }} />
 
-            {/*<ShowTicketType {...{tickets, buttonsDisabled, }} /> */}
-            <div className="d-flex justify-content-center">
-                <table>
-                    <colgroup>
-                        <col style={{ width: '120px' }} />
-                        <col style={{ width: '40px' }} />
-                        <col style={{ width: '30px' }} />
-                        <col style={{ width: '40px' }} />
-                    </colgroup>
-                    <tbody>
-                        <tr>
-                            <td>Vuxen</td>
-                            <td className="d-flex justify-content-center"><Button style={{ display: 'none' }} disabled>–</Button></td>
-                            <td className="text-center">{tickets.ordinary}</td>
-                            <td className="d-flex justify-content-center"><Button style={{ display: 'none' }} disabled>+</Button></td>
-                        </tr>
-                        <tr>
-                            <td>Barn</td>
-                            <td className="d-flex justify-content-center"><Button onClick={() => decreaseTicketCount('barn')} variant="danger" disabled={buttonsDisabled}>–</Button></td>
-                            <td className="text-center">{tickets.child}</td>
-                            <td className="d-flex justify-content-center"><Button onClick={() => increaseTicketCount('barn')} variant="primary" disabled={buttonsDisabled}>+</Button></td>
-                        </tr>
-                        <tr>
-                            <td>Pensionär</td>
-                            <td className="d-flex justify-content-center"><Button onClick={() => decreaseTicketCount('pensionar')} variant="danger" disabled={buttonsDisabled}>–</Button></td>
-                            <td className="text-center">{tickets.pensioner}</td>
-                            <td className="d-flex justify-content-center"><Button onClick={() => increaseTicketCount('pensionar')} variant="primary" disabled={buttonsDisabled}>+</Button></td>
-                        </tr>
-                    </tbody>
-                </table>
-            </div>
+            <ShowTicketType {...{tickets, buttonsDisabled, setTickets}} /> 
+
             <Row>
                 <Col className="d-flex justify-content-center mt-3">
                     <span style={{ fontSize: '22px' }}>Att betala: {summa} kr</span>
